+++ src/library/DIPS.Mobile.UI/Components/Saving/SaveView/SaveView.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.API.Vibration;
 using DIPS.Mobile.UI.Components.CheckBoxes;
 using Colors = Microsoft.Maui.Graphics.Colors;
 using Label = DIPS.Mobile.UI.Components.Labels.Label;
@@ -8,17 +9,17 @@ namespace DIPS.Mobile.UI.Components.Saving.SaveView;
 public partial class SaveView : ContentView
 {
     private readonly Label m_stateLabel;
+    private readonly FilledCheckBox m_filledCheckBox;
 
     public SaveView()
     {
-        var filledCheckBox = new FilledCheckBox
-        {
-            VerticalOptions = LayoutOptions.Center,
-        };
-        
-        filledCheckBox.SetBinding(FilledCheckBox.IsCheckedProperty, new Binding(nameof(IsSavingCompleted), source: this));
-        filledCheckBox.SetBinding(FilledCheckBox.IsProgressingProperty, new Binding(nameof(IsSaving), source: this));
-        filledCheckBox.SetBinding(FilledCheckBox.CompletedCommandProperty, new Binding(nameof(SavingCompletedCommand), source: this));
+        m_filledCheckBox = new FilledCheckBox {VerticalOptions = LayoutOptions.Center,};
+
+        m_filledCheckBox.SetBinding(FilledCheckBox.IsCheckedProperty,
+            new Binding(nameof(IsSavingCompleted), source: this));
+        m_filledCheckBox.SetBinding(FilledCheckBox.IsProgressingProperty, new Binding(nameof(IsSaving), source: this));
+        m_filledCheckBox.SetBinding(FilledCheckBox.CompletedCommandProperty,
+            new Binding(nameof(SavingCompletedCommand), source: this));
 
         m_stateLabel = new Label
         {
@@ -34,10 +35,13 @@ public partial class SaveView : ContentView
             HorizontalOptions = LayoutOptions.Center,
             VerticalOptions = LayoutOptions.Center,
             Spacing = Sizes.GetSize(SizeName.size_12),
-            Children = { filledCheckBox, m_stateLabel }
+            Children = {m_filledCheckBox, m_stateLabel}
         };
+    }
 
-        Content = content;
+    private void OnAnimationFinished(object? sender, EventArgs e)
+    {
+        VibrationService.SelectionChanged();
     }
 
     protected override void OnHandlerChanged()
@@ -60,6 +64,7 @@ public partial class SaveView : ContentView
         if (newValue is true)
         {
             saveView.SetSavingCompletedText();
+            VibrationService.SelectionChanged();
         }
     }